\name{hasseDiagram-package}
\alias{hasseDiagram-package}
\alias{hasseDiagram}
\docType{package}
\title{
Drawing Hasse diagram
}
\description{
Drawing Hasse diagram - visualization of transitive reduction of a finite partially ordered set.
}
\details{
\tabular{ll}{
Package: \tab hasseDiagram\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2014-05-20\cr
License: \tab MIT\cr
}
}
\author{
Krzysztof Ciomek

Maintainer: Krzysztof Ciomek <k.ciomek@gmail.com>
}
\keyword{ hasse diagram finite partially ordered set }
\seealso{
\code{\link{hasse}}
}
